code refactor & fix

+++ e-commerce/routes/web.php
@@ -45,6 +45,7 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::get('users/edit/{id}', [UserController::class, 'edit'])->name('user.edit');
     Route::post('users/edit/{id}', [UserController::class, 'updateCart'])->name('user.update');
     Route::get('/profile/{id}', [UserController::class, 'showProfile'])->name('profile');
+    Route::get('/mypost/{user_id}', [UserController::class, 'getMyPost'])->name('mypost.show');
 
     Route::post('like', [ProductController::class, 'like'])->name('like');
     Route::delete('like', [ProductController::class, 'unlike'])->name('unlike');
@@ -71,9 +72,9 @@ Route::get('/export/categories', [AdminCategoryController::class, 'exportCategor
 Route::post('import/categories', [AdminCategoryController::class, 'importCategory'])->name('categories.import');
 
 Route::get('product/lists/show', [AdminProductController::class, 'getAllProducts'])->name('products.lists');
-Route::get('product/edit/{product}', [AdminProductController::class, 'edit'])->name('product.edit');
-Route::post('product/edit/{product}', [AdminProductController::class, 'update'])->name('update.product');
-Route::delete('/product/{product}', [AdminProductController::class, 'destroy'])->name('product.remove');
+Route::get('admin/edit/{product}', [AdminProductController::class, 'edit'])->name('edit.product');
+Route::post('admin/edit/{product}', [AdminProductController::class, 'update'])->name('update.product');
+Route::delete('admin/product/{product}', [AdminProductController::class, 'destroy'])->name('product.remove');
 
 Route::get('/admin/dashboard', function () {
